requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
pyasn1==0.5.1
pytest==7.4.3
python-benedict==0.33.1
python-fsutil==0.13.0
python-slugify==8.0.1
requests==2.31.0
rsa==4.9
text-unidecode==1.3
urllib3==2.1.0
Werkzeug==3.0.1
